requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
build==1.3.0
    # via pip-tools
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
click==8.3.0
    # via pip-tools
distro==1.9.0
    # via openai
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.11.0
    # via openai
openai==2.2.0
    # via -r requirements.in
packaging==25.0
    # via build
pip-tools==7.5.1
    # via -r requirements.in
pydantic==2.11.10
    # via openai
pydantic-core==2.33.2
    # via pydantic
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.1.1
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
